Refactored aggregate.
